+++ ChangeLog.md
@@ -13,6 +13,7 @@
     - Field.configure() becomes a reactive data source
     - Accept dt_data extended definition
     - Accept and ignore null or empty declarations (todo #1)
+    - Accept (and flat-ize) array declarations inside of array declarations, and recursively (todo #2)
 
 ### 1.1.0